requirements.txt
# Configuration and data handling
pyyaml
python-dotenv>=1.0.0

# Async and concurrency
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1

# HTTP and web requests
aiohttp>=3.9.0
requests>=2.31.0

# Database support (optional)
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Logging and monitoring
structlog>=23.2.0
colorama>=0.4.6

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# API integrations
openai>=1.3.0
anthropic>=0.7.0

# Web scraping (for bug hunter agent)
beautifulsoup4>=4.12.0
selenium>=4.15.0
scrapy>=2.11.0

# Email and notifications
smtplib2>=0.2.1
sendgrid>=6.10.0

# File processing
python-magic>=0.4.27
openpyxl>=3.1.0

# Security
cryptography>=41.0.0
bcrypt>=4.1.0

# Utilities
click>=8.1.0
tqdm>=4.66.0
schedule>=1.2.0

# Optional: Machine Learning
# scikit-learn>=1.3.0
# tensorflow>=2.13.0
# transformers>=4.35.0

# Optional: Database drivers
# psycopg2-binary>=2.9.7  # PostgreSQL
# pymongo>=4.5.0          # MongoDB
# redis>=5.0.0            # Redis

# Optional: Message queues
# celery>=5.3.0           # Task queue
# kombu>=5.3.0            # Message library

# Optional: Monitoring and metrics
# prometheus-client>=0.18.0
# grafana-api>=1.0.3

# Development tools (optional)
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Documentation (optional)
# sphinx>=7.2.0
# sphinx-rtd-theme>=1.3.0